Remove redundant check for successful installation
This commit removes a redundant check in the nodemanager that would
re-check the crontab install flag from `nodeman.cfg` and try to
update the crontab and flag once more.

However, setting the crontab and the flag is `seattleinstaller`'s
job, see SeattleTestbed/installer-packaging. It also reports errors.
There is no point in the nodemanager duplicating half of this
functionality.