Amend `nminit.py` in response to #128
`nminit.py` creates the nodeman config for unit-testing the NM module.
This commit makes it set the `seattle_installed` flag in the config to
`True`, emulating a full and proper install. Otherwise, the patch for
issue #128 keeps the nodemanager from starting.

The commit also fixes a syntax error in the nodemanger ports list. This
error is masked by the roundabout way in which `nminit` is `Popen`ed
during unit testing. (See #125 for a fix of this call problem.)

Lastly, it adds a new log message to indicate that the unit test
initialization is done, just before the actual test cases are run.